+++ code/NCheck.Test/CheckerFactoryFixture.cs
@@ -1,0 +1,9 @@
+﻿namespace NCheck.Test
+{
+    using NUnit.Framework;
+
+    [TestFixture]
+    public class CheckerFactoryFixture : Fixture
+    {
+    }
+}