+++ code/NCheck.Test/CheckerFixture.cs
@@ -39,28 +39,64 @@
         }
 
         [Test]
+        public void ValidateCompareValueWithCustomComparer()
+        {
+            var expected = new Simple { Id = 1, Value = 1 };
+            var candidate = new Simple { Id = 1, Value = 1.0005 };
+
+            Check(expected, candidate);
+        }
+
+        [Test]
+        public void ValidateCompareValueWithCustomComparerFault()
+        {
+            var expected = new Simple { Id = 1, Value = 1 };
+            var candidate = new Simple { Id = 1, Value = 1.1 };
+
+            CheckFault(expected, candidate, "Simple.Value", "1", "1.1");
+        }
+
+        [Test]
         public void ValidateCompareById()
         {
-            var p1 = new Parent { Id = 1, Name = "A" };
-            var expected = new Child { Id = 1, Name = "Child", Parent = p1 };
+            var p1 = new Parent {Id = 1, Name = "A"};
+            var expected = new Child {Id = 1, Name = "Child", Parent = p1};
             p1.Favourite = expected;
 
-            var p2 = new Parent { Id = 1, Name = "A" };
-            var candidate = new Child { Id = 1, Name = "Child", Parent = p2 };
+            var p2 = new Parent {Id = 1, Name = "A"};
+            var candidate = new Child {Id = 1, Name = "Child", Parent = p2};
             p2.Favourite = candidate;
 
             Check(expected, candidate);
         }
 
         [Test]
-        public void ValidateCompareByIdMessage()
+        public void ValidateCompareByIdWithNullIdentityChecker()
         {
-            var p1 = new Parent { Id = 1, Name = "A" };
-            var expected = new Child { Id = 1, Name = "Child", Parent = p1 };
+            var p1 = new Parent {Id = 1, Name = "A"};
+            var expected = new Child {Id = 1, Name = "Child", Parent = p1};
             p1.Favourite = expected;
 
-            var p2 = new Parent { Id = 2, Name = "A" };
-            var candidate = new Child { Id = 1, Name = "Child", Parent = p2 };
+            var p2 = new Parent {Id = 1, Name = "A"};
+            var candidate = new Child {Id = 1, Name = "Child", Parent = p2};
+            p2.Favourite = candidate;
+
+            // NB Need to initialize the checker factory
+            var x = CheckerFactory;
+            PropertyCheck.IdentityChecker = null;
+
+            CheckFault(expected, candidate, "No IdentityChecker assigned, cannot perform Id check");
+        }
+
+        [Test]
+        public void ValidateCompareByIdMessage()
+        {
+            var p1 = new Parent {Id = 1, Name = "A"};
+            var expected = new Child {Id = 1, Name = "Child", Parent = p1};
+            p1.Favourite = expected;
+
+            var p2 = new Parent {Id = 2, Name = "A"};
+            var candidate = new Child {Id = 1, Name = "Child", Parent = p2};
             p2.Favourite = candidate;
 
             CheckFault(expected, candidate, "Child.Parent.Id", 1, 2);
@@ -69,8 +105,8 @@
         [Test]
         public void ValidateExcludeProperty()
         {
-            var expected = new Parent { Id = 1, Name = "A", Ignore = "A", Another = 1 };
-            var candidate = new Parent { Id = 1, Name = "A", Ignore = "B", Another = 2 };
+            var expected = new Parent {Id = 1, Name = "A", Ignore = "A", Another = 1};
+            var candidate = new Parent {Id = 1, Name = "A", Ignore = "B", Another = 2};
 
             Check(expected, candidate);
         }
@@ -78,12 +114,12 @@
         [Test]
         public void ValidatePropertyMessageInChildEntity()
         {
-            var expected = new Parent { Id = 1, Name = "A" };
-            var c1 = new Child { Id = 1, Name = "Child", Parent = expected };
+            var expected = new Parent {Id = 1, Name = "A"};
+            var c1 = new Child {Id = 1, Name = "Child", Parent = expected};
             expected.Favourite = c1;
 
-            var candidate = new Parent { Id = 1, Name = "A" };
-            var c2 = new Child { Id = 2, Name = "Child", Parent = candidate };
+            var candidate = new Parent {Id = 1, Name = "A"};
+            var c2 = new Child {Id = 2, Name = "Child", Parent = candidate};
             candidate.Favourite = c2;
 
             CheckFault(expected, candidate, "Parent.Favourite.Id", 1, 2);
@@ -93,8 +129,8 @@
         public void ValidateEntityWithList()
         {
             var c1 = new Child();
-            var expected = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
-            var candidate = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
+            var expected = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
+            var candidate = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
 
             expected.Children.Add(c1);
             candidate.Children.Add(c1);
@@ -106,8 +142,8 @@
         public void ValidateListExplicitly()
         {
             var c1 = new Child();
-            var expected = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
-            var candidate = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
+            var expected = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
+            var candidate = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
 
             expected.Children.Add(c1);
             candidate.Children.Add(c1);
@@ -118,8 +154,8 @@
         [Test]
         public void ValidateNullIIdentifiable()
         {
-            var expected = new Child { Id = 1, Name = "Child" };
-            var candidate = new Child { Id = 1, Name = "Child" };
+            var expected = new Child {Id = 1, Name = "Child"};
+            var candidate = new Child {Id = 1, Name = "Child"};
 
             Check(expected, candidate);
         }
@@ -127,8 +163,8 @@
         [Test]
         public void ValidateWhenBothCollectionPropertyAreNull()
         {
-            var expected = new Parent { Id = 1, Name = "A" };
-            var candidate = new Parent { Id = 1, Name = "A" };
+            var expected = new Parent {Id = 1, Name = "A"};
+            var candidate = new Parent {Id = 1, Name = "A"};
 
             Check(expected, candidate);
         }
@@ -136,8 +172,8 @@
         [Test]
         public void ValidateMessageWhenCandidateCollectionPropertyIsNull()
         {
-            var expected = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
-            var candidate = new Parent { Id = 1, Name = "A" };
+            var expected = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
+            var candidate = new Parent {Id = 1, Name = "A"};
 
             CheckFault(expected, candidate, "Parent.Children", "not null", "null");
         }
@@ -145,8 +181,8 @@
         [Test]
         public void ValidateMessageWhenExpectedCollectionPropertyIsNull()
         {
-            var expected = new Parent { Id = 1, Name = "A" };
-            var candidate = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
+            var expected = new Parent {Id = 1, Name = "A"};
+            var candidate = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
 
             CheckFault(expected, candidate, "Parent.Children", "null", "not null");
         }
@@ -155,8 +191,8 @@
         public void ValidateMessageWhenCandidateCollectionCountIsDifferent()
         {
             var c1 = new Child();
-            var expected = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
-            var candidate = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
+            var expected = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
+            var candidate = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
 
             expected.Children.Add(c1);
 
@@ -167,8 +203,8 @@
         public void ValidateMessageWhenExpectedCollectionCountIsDifferent()
         {
             var c1 = new Child();
-            var expected = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
-            var candidate = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
+            var expected = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
+            var candidate = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
 
             candidate.Children.Add(c1);
 
@@ -178,12 +214,12 @@
         [Test]
         public void ValidateMessageCollectionElementIsDifferent()
         {
-            var expected = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
-            var candidate = new Parent { Id = 1, Name = "A", Children = new List<Child>() };
+            var expected = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
+            var candidate = new Parent {Id = 1, Name = "A", Children = new List<Child>()};
 
-            var c1 = new Child { Id = 1 };
+            var c1 = new Child {Id = 1};
             expected.Children.Add(c1);
-            var c2 = new Child { Id = 2 };
+            var c2 = new Child {Id = 2};
             candidate.Children.Add(c2);
 
             CheckFault(expected, candidate, "Parent.Children[0].Id", 1, 2);
@@ -206,13 +242,13 @@
         [Test]
         public void NoIIdentifiableCheckerRegistered()
         {
-            var cf = (CheckerFactory)CheckerFactory;
+            var cf = (CheckerFactory) CheckerFactory;
             cf.Clear();
             PropertyCheck.IdentityChecker = null;
 
-            var p1 = new Parent { Id = 1, Name = "A" };
-            var expected = new Child { Id = 1, Name = "Child", Parent = p1 };
-            var candidate = new Child { Id = 2, Name = "Child", Parent = p1 };
+            var p1 = new Parent {Id = 1, Name = "A"};
+            var expected = new Child {Id = 1, Name = "Child", Parent = p1};
+            var candidate = new Child {Id = 2, Name = "Child", Parent = p1};
 
             CheckFault(expected, candidate, "No IdentityChecker assigned, cannot perform Id check");
         }
@@ -220,39 +256,15 @@
         [Test]
         public void NullBuilderMeansNoAutoRegistration()
         {
-            var cf = (CheckerFactory)CheckerFactory;
+            var cf = (CheckerFactory) CheckerFactory;
             cf.Clear();
             cf.Builder = new NullCheckerBuilder();
 
-            var p1 = new Parent { Id = 1, Name = "A" };
-            var expected = new Child { Id = 1, Name = "Child", Parent = p1 };
-            var candidate = new Child { Id = 2, Name = "Child", Parent = p1 };
+            var p1 = new Parent {Id = 1, Name = "A"};
+            var expected = new Child {Id = 1, Name = "Child", Parent = p1};
+            var candidate = new Child {Id = 2, Name = "Child", Parent = p1};
 
             CheckFault(expected, candidate, "No checker registered for NCheck.Test.Child");
-        }
-
-        private void CheckFault<T>(T expected, T candidate, string name, object expectedValue, object actualValue)
-        {
-            const string messageFormat = "{0}: Expected:<{1}>. Actual:<{2}>";
-
-            var message = string.Format(messageFormat, name, expectedValue, actualValue);
-
-            CheckFault(expected, candidate, message);
-        }
-
-        private void CheckFault<T>(T expected, T candidate, string faultMessage)
-        {
-            try
-            {
-                Check(expected, candidate);
-            }
-            catch (Exception ex)
-            {
-                Assert.AreEqual(faultMessage, ex.Message);
-                return;
-            }
-
-            Assert.Fail("No exception, expected: " + faultMessage);
         }
     }
 }