+++ code/NCheck/CheckerExtensions.cs
@@ -5,7 +5,7 @@
     using System.Collections.Generic;
     using System.Reflection;
 
-    using NCheck.Checking;
+    using Checking;
 
     /// <summary>
     /// Extension methods for the checkers.
@@ -47,61 +47,93 @@
         }
 
         /// <summary>
-        /// Sets the default behaviour for a <see cref="ITypeCompareTargeter"/>
+        /// Sets the default behaviour for a <see cref="IConventions{Type}"/>
         /// </summary>
-        /// <param name="targeter"></param>
-        public static void InitializeTypeCompareTargeter(this ITypeCompareTargeter targeter)
+        /// <param name="conventions"></param>
+        public static void InitializeTypeConventions(this IConventions<Type> conventions)
         {
-            targeter.Register(ReferenceTypeAsEntity);
-            targeter.Register(NonNullableStructsAsEntity);
-            targeter.Register(EnumerableAsCollection);
-            targeter.Register(IsEnum);
+            conventions.Clear();
+            conventions.CompareTarget.InitializeTypeCompareTargetConventions();
+        }
+
+        /// <summary>
+        /// Sets the default behaviour for a <see cref="IConventions{Type, CompareTarget}"/>
+        /// </summary>
+        /// <param name="conventions"></param>
+        public static void InitializeTypeCompareTargetConventions(this IConventions<Type, CompareTarget> conventions)
+        {
+            conventions.Register(ReferenceTypeAsEntity, CompareTarget.Entity);
+            conventions.Register(NonNullableStructsAsEntity, CompareTarget.Entity);
+            conventions.Register(EnumerableAsCollection, CompareTarget.Collection);
+            conventions.Register(IsEnum, CompareTarget.Value);
 
             // Put the default overrides in
-            targeter.Register(typeof(Guid), CompareTarget.Value);
-            targeter.Register(typeof(String), CompareTarget.Value);
-            targeter.Register(typeof(Decimal), CompareTarget.Value);
-            targeter.Register(typeof(Single), CompareTarget.Value);
-            targeter.Register(typeof(Double), CompareTarget.Value);
-            targeter.Register(typeof(DateTime), CompareTarget.Value);
-            targeter.Register(typeof(DateTimeOffset), CompareTarget.Value);
-            targeter.Register(typeof(TimeSpan), CompareTarget.Value);
-            targeter.Register(typeof(TimeZone), CompareTarget.Value);
-            targeter.Register(typeof(TimeZoneInfo), CompareTarget.Value);
+            conventions.Register(typeof(Guid), CompareTarget.Value);
+            conventions.Register(typeof(String), CompareTarget.Value);
+            conventions.Register(typeof(Decimal), CompareTarget.Value);
+            conventions.Register(typeof(Single), CompareTarget.Value);
+            conventions.Register(typeof(Double), CompareTarget.Value);
+            conventions.Register(typeof(DateTime), CompareTarget.Value);
+            conventions.Register(typeof(DateTimeOffset), CompareTarget.Value);
+            conventions.Register(typeof(TimeSpan), CompareTarget.Value);
+            conventions.Register(typeof(TimeZone), CompareTarget.Value);
+            conventions.Register(typeof(TimeZoneInfo), CompareTarget.Value);
 
             // Type is abstract, is RuntimeType at r/t which is internal so can't handle easily
-            targeter.Register(typeof(Type), CompareTarget.Value);            
+            conventions.Register(typeof(Type), CompareTarget.Value);      
         }
 
-        private static CompareTarget IsEnum(Type type)
+        /// <summary>
+        /// Converts a strongly typed function into a comparer convention
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="func"></param>
+        public static Func<object, object, bool> ToComparerConvention<T>(this Func<T, T, bool> func)
         {
-            return type.IsEnum ? CompareTarget.Value : CompareTarget.Unknown;
+            // NB The cast could fail at runtime and this will also cause boxing.
+            return (x, y) => func((T) x, (T) y);
         }
 
-        private static CompareTarget NonNullableStructsAsEntity(Type type)
+        /// <summary>
+        /// Register an explicit <see cref="CompareTarget"/> to use for a type.
+        /// </summary>
+        /// <param name="conventions">Comparer to register against</param>
+        /// <param name="type">Type to use</param>
+        /// <param name="target">CompareTarget to use</param>
+        public static void Register(this IConventions<Type, CompareTarget> conventions, Type type, CompareTarget target)
+        {
+            conventions.Register(x => x.FullName == type.FullName, target);
+        }
+
+        private static bool IsEnum(Type type)
+        {
+            return type.IsEnum;
+        }
+
+        private static bool NonNullableStructsAsEntity(Type type)
         {
             if (!type.IsValueType)
             {
-                return CompareTarget.Unknown;
+                return false;
             }
 
             // Trying to make sure we process structs as entities, but exclude nullables
             if (!type.IsPrimitive && !type.FullName.StartsWith("System.Nullable"))
             {
-                return CompareTarget.Entity;
+                return true;
             }
 
-            return CompareTarget.Unknown;
+            return false;
         }
 
-        private static CompareTarget ReferenceTypeAsEntity(Type type)
+        private static bool ReferenceTypeAsEntity(Type type)
         {
-            return !type.IsValueType ? CompareTarget.Entity : CompareTarget.Unknown;
+            return !type.IsValueType;
         }
 
-        private static CompareTarget EnumerableAsCollection(Type type)
+        private static bool EnumerableAsCollection(Type type)
         {
-            return typeof(IEnumerable).IsAssignableFrom(type) ? CompareTarget.Collection : CompareTarget.Unknown;
+            return typeof(IEnumerable).IsAssignableFrom(type);
         }
     }
 }