+++ code/NCheck/Checking/PropertyConventions.cs
@@ -1,0 +1,11 @@
+﻿namespace NCheck.Checking
+{
+    using System.Reflection;
+
+    /// <summary>
+    /// Implements property conventions
+    /// </summary>
+    public class PropertyConventions : Conventions<PropertyInfo>
+    {
+    }
+}